add skills page and list
